docs: add ARACA report
