+++ ARACA_REPORT_20250731-070132.md
@@ -1,0 +1,26 @@
+# ARACA Report - 2025-07-31T07:01:32Z
+
+## Objectives and Constraints
+- Refactor codebase for improved structure.
+- Preserve Eleventy/Nunjucks/Tailwind stack.
+- Disable footnote bottom forcing and fix popup on mobile.
+
+## Architecture Overview
+- **lib/markdown/** – new module directory housing footnote logic (`footnotes.js`), inline macros (`inlineMacros.js`), link tweaks (`links.js`), and aggregator (`index.js`).
+- **src/assets/css/tailwind.css** – adjusted mobile styles for annotation popups.
+- **tests** – updated footnote tests verifying absence of bottom footnote block.
+
+## Implementation Summary
+- **Modularization**: split former `markdown-extensions.js` into multiple focused files under `lib/markdown/` and adjusted imports accordingly.
+- **Configuration Hygiene**: added `disableFootnoteTail` to remove unwanted footer injection.
+- **Type/Contract Strengthening**: added guards when collecting footnote tokens and documented functions via JSDoc.
+- **Naming/Locality**: centralized extension list via `lib/markdown/index.js` improving clarity.
+- **Performance Trivial**: refined mobile popup CSS to use `inset-x-4` for responsive centering.
+
+## Validation Results
+- `npm test` – all tests pass including new check for disabled footnote tail【2d6b4a†L1-L19】.
+- `npm run build` – Eleventy build succeeds generating 34 files despite wiki warnings【bf060b†L1-L31】.
+
+## Next Steps
+- Monitor for any edge cases in footnote parsing where definitions are missing.
+- Evaluate further modularisation of build steps or CSS.
