+++ lib/markdown/footnotes.js
@@ -1,12 +1,14 @@
-// Custom markdown-it extensions used by Eleventy
 /**
- * Collection of markdown-it extension functions for footnotes, embeds and links.
- * @module markdownExtensions
+ * Footnote rendering tweaks for markdown-it.
+ * - Disables footer output via footnote_tail.
+ * - Renders inline popover markup for each reference.
+ * - Captures definition tokens for later reuse.
+ * @module footnotes
  */
 
 /**
- * Enhance the default footnote output to support local display and hybrid blocks.
- * @param {import('markdown-it')} md - markdown-it instance to mutate
+ * Enhance the default footnote output to support local display.
+ * @param {import('markdown-it')} md markdown-it instance
  */
 function hybridFootnoteDefinitions(md) {
   md.renderer.rules.footnote_reference_open = (tokens, idx, _opts, env) => {
@@ -24,11 +26,9 @@ function hybridFootnoteDefinitions(md) {
 
   md.renderer.rules.footnote_block_open = () => '';
   md.renderer.rules.footnote_block_close = () => '';
-
   md.renderer.rules.footnote_open = () => '';
   md.renderer.rules.footnote_close = () => '';
   md.renderer.rules.footnote_anchor = () => '';
-
   md.renderer.rules.footnote_body_open = () => '';
   md.renderer.rules.footnote_body_close = () => '';
   md.renderer.rules.footnote_caption = () => '';
@@ -41,7 +41,7 @@ function hybridFootnoteDefinitions(md) {
 
 /**
  * Replace footnote references with popover-enabled anchors.
- * @param {import('markdown-it')} md - markdown-it instance to mutate
+ * @param {import('markdown-it')} md markdown-it instance
  */
 function footnotePopover(md) {
   const originalFootnoteRef = md.renderer.rules.footnote_ref;
@@ -53,7 +53,6 @@ function footnotePopover(md) {
     if (!Array.isArray(list) || !Array.isArray(list[id]?.tokens)) {
       return originalFootnoteRef(tokens, idx, options, env, self);
     }
-
     const n = id + 1;
     let defHtml = '';
     try {
@@ -71,47 +70,8 @@ function footnotePopover(md) {
 }
 
 /**
- * Helper to define simple inline macros.
- * @param {string} name - macro name
- * @param {string} after - rule to insert after
- * @param {(value:string)=>string} toHtml - HTML generator
- */
-const inlineMacro = (name, after, toHtml) => md => {
-  md.inline.ruler.after(after, name, (state, silent) => {
-    const m = state.src.slice(state.pos).match(new RegExp(`^@${name}\\(([^)]+)\\)`));
-    if (!m) return false;
-    if (!silent) state.push({ type: 'html_inline', content: toHtml(m[1]) });
-    state.pos += m[0].length;
-    return true;
-  });
-};
-
-/** Inline audio embedding macro */
-const audioEmbed = inlineMacro('audio', 'emphasis', src => `<audio controls class="audio-embed" src="${src}"></audio>`);
-
-/** Inline QR-code embedding macro */
-const qrEmbed = inlineMacro('qr', 'audio', s => {
-  const src = encodeURIComponent(s);
-  return `<img class="qr-code" src="https://api.qrserver.com/v1/create-qr-code/?size=150x150&data=${src}" alt="QR code">`;
-});
-
-/**
- * Add an external-link class and arrow to outbound links.
- * @param {import('markdown-it')} md - markdown-it instance
- */
-function externalLinks(md) {
-  const base = md.renderer.rules.link_open || ((t, i, o, e, s) => s.renderToken(t, i, o));
-  md.renderer.rules.link_open = (tokens, idx, options, env, self) => {
-    tokens[idx].attrJoin('class', 'external-link');
-    const nxt = tokens[idx + 1];
-    if (nxt?.type === 'text') nxt.content = '↗ source';
-    return base(tokens, idx, options, env, self);
-  };
-}
-
-/**
  * Capture footnote definition tokens when footnote_tail is disabled.
- * @param {import('markdown-it')} md - markdown-it instance
+ * @param {import('markdown-it')} md markdown-it instance
  */
 function collectFootnoteTokens(md) {
   md.core.ruler.after('inline', 'collect_footnote_tokens', state => {
@@ -125,28 +85,26 @@ function collectFootnoteTokens(md) {
       } else if (tok.type === 'footnote_reference_close') {
         const entry = stack.pop();
         if (!entry) return;
+        const target = env.footnotes.list[entry.id];
+        if (!target) return;
         const tokens = state.tokens.slice(entry.start + 1, idx);
-        env.footnotes.list[entry.id].tokens = tokens;
+        target.tokens = tokens;
       }
     });
   });
 }
 
-const mdItExtensions = [
-  hybridFootnoteDefinitions,
-  footnotePopover,
-  collectFootnoteTokens,
-  audioEmbed,
-  qrEmbed,
-  externalLinks
-];
+/**
+ * Disable markdown-it-footnote's tail injection behavior.
+ * @param {import('markdown-it')} md markdown-it instance
+ */
+function disableFootnoteTail(md) {
+  md.core.ruler.disable('footnote_tail');
+}
 
 module.exports = {
   hybridFootnoteDefinitions,
   footnotePopover,
   collectFootnoteTokens,
-  audioEmbed,
-  qrEmbed,
-  externalLinks,
-  mdItExtensions
+  disableFootnoteTail
 };
