+++ lib/markdown/inlineMacros.js
@@ -1,0 +1,26 @@
+/**
+ * Helper to define simple inline macros.
+ * @param {string} name - macro name
+ * @param {string} after - rule to insert after
+ * @param {(value:string)=>string} toHtml - HTML generator
+ */
+const inlineMacro = (name, after, toHtml) => md => {
+  md.inline.ruler.after(after, name, (state, silent) => {
+    const m = state.src.slice(state.pos).match(new RegExp(`^@${name}\\(([^)]+)\\)`));
+    if (!m) return false;
+    if (!silent) state.push({ type: 'html_inline', content: toHtml(m[1]) });
+    state.pos += m[0].length;
+    return true;
+  });
+};
+
+/** Inline audio embedding macro */
+const audioEmbed = inlineMacro('audio', 'emphasis', src => `<audio controls class="audio-embed" src="${src}"></audio>`);
+
+/** Inline QR-code embedding macro */
+const qrEmbed = inlineMacro('qr', 'audio', s => {
+  const src = encodeURIComponent(s);
+  return `<img class="qr-code" src="https://api.qrserver.com/v1/create-qr-code/?size=150x150&data=${src}" alt="QR code">`;
+});
+
+module.exports = { inlineMacro, audioEmbed, qrEmbed };
