+++ lib/markdown/links.js
@@ -1,0 +1,15 @@
+/**
+ * Add an external-link class and arrow to outbound links.
+ * @param {import('markdown-it')} md - markdown-it instance
+ */
+function externalLinks(md) {
+  const base = md.renderer.rules.link_open || ((t, i, o, e, s) => s.renderToken(t, i, o));
+  md.renderer.rules.link_open = (tokens, idx, options, env, self) => {
+    tokens[idx].attrJoin('class', 'external-link');
+    const nxt = tokens[idx + 1];
+    if (nxt?.type === 'text') nxt.content = '↗ source';
+    return base(tokens, idx, options, env, self);
+  };
+}
+
+module.exports = { externalLinks };
